Updated to match data.sql
Updated many classes and data.sql to match the relationship diagram. It runs but still has issues concerning ISBN and FK constraints. Will try to fix later.

+++ bookstore/src/main/java/app/bookstore/BookRepo.java
@@ -2,7 +2,7 @@ package app.bookstore;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface BookRepo extends CrudRepository<Book, Integer>
+public interface BookRepo extends CrudRepository<Book, Long>
 {
 //	Book findByISBN(long ISBN);
 //
